// ombi-api/model/ombi-store-entities-requests-movie-requests.ts
export interface OmbiStoreEntitiesRequestsMovieRequests {
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    theMovieDbId?: number;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    issueId?: number;
    /**
     * 
     * @type {Array<OmbiStoreEntitiesRequestsIssues>}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    issues?: Array<OmbiStoreEntitiesRequestsIssues>;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    subscribed?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    showSubscribe?: boolean;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    rootPathOverride?: number;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    qualityOverride?: number;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    langCode?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    imdbId?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    overview?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    posterPath?: string;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    releaseDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    digitalReleaseDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    status?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    background?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    released?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    digitalRelease?: boolean;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    title?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    approved?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    markedAsApproved?: Date;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    requestedDate?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    available?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    markedAsAvailable?: Date;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    requestedUserId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    denied?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    markedAsDenied?: Date;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    deniedReason?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    requestType?: OmbiStoreEntitiesRequestsMovieRequestsRequestTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    requestedByAlias?: string;
    /**
     * 
     * @type {OmbiStoreEntitiesOmbiUser}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    requestedUser?: OmbiStoreEntitiesOmbiUser;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    canApprove?: boolean;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsMovieRequests
     */
    id?: number;
}

/**
    * @export
    * @enum {string}
    */
export enum OmbiStoreEntitiesRequestsMovieRequestsRequestTypeEnum {
    TvShow = 'TvShow',
    Movie = 'Movie',
    Album = 'Album'
}